Better consistency between inlined values and binding values
