+++ squidb-addons/squidb-sqlite-bindings/src/com/yahoo/squidb/sqlitebindings/SQLiteBindingsCursorFactory.java
@@ -7,7 +7,7 @@ package com.yahoo.squidb.sqlitebindings;
 
 import android.database.Cursor;
 
-import com.yahoo.squidb.data.adapter.SquidCursorFactory;
+import com.yahoo.squidb.sql.SqlUtils;
 
 import org.sqlite.database.sqlite.SQLiteCursor;
 import org.sqlite.database.sqlite.SQLiteCursorDriver;
@@ -41,7 +41,7 @@ public class SQLiteBindingsCursorFactory implements CursorFactory {
             return;
         }
         for (int i = 1; i <= sqlArgs.length; i++) {
-            Object arg = SquidCursorFactory.resolveArgReferences(sqlArgs[i - 1]);
+            Object arg = SqlUtils.resolveArgReferences(sqlArgs[i - 1]);
             bindObjectToProgram(program, i, arg);
         }
     }
